added docstrings and updated tests with interpolated bathy mesh
