+++ proteus/tests/test_meshtools.py
@@ -4,8 +4,11 @@ import numpy.testing as npt
 import numpy as np
 from nose.tools import eq_, ok_
 import math
+import xml.etree.ElementTree as ElementTree
 from proteus.EGeometry import (EVec,
                                ETen)
+from proteus.Archiver import XdmfArchive
+from proteus.Domain import InterpolatedBathymetryDomain
 from proteus.MeshTools import (Node,
                                Element,
                                Edge,
@@ -32,7 +35,9 @@ from proteus.MeshTools import (Node,
                                getMeshIntersections,
                                MultilevelEdgeMesh,
                                MultilevelTriangularMesh,
-                               MultilevelTetrahedralMesh)
+                               MultilevelTetrahedralMesh,
+                               MultilevelHexahedralMesh,
+                               InterpolatedBathymetryMesh)
 
 comm = Comm.init()
 Profiling.procID = comm.rank()
@@ -553,6 +558,12 @@ def test_MultilevelTriangularMesh():
             ok_((elementChildren[l] == mlMesh.elementChildrenArrayList[l]).all())
             ok_((elementChildrenOffsets[l] == mlMesh.elementChildrenOffsetsList[l]).all())
         ok_((elementParents[l] == mlMesh.elementParentsArrayList[l]).all())
+    mlMesh = MultilevelTriangularMesh(3,3,1,refinementLevels=n,triangleFlag=1)
+    mlMesh = MultilevelTriangularMesh(3,3,1,refinementLevels=n,triangleFlag=2)
+    mlMesh2 = MultilevelTriangularMesh(0,0,0,skipInit=True)
+    mlMesh2.generateFromExistingCoarseMesh(mlMesh.meshList[0],
+                                           refinementLevels=n)
+
 
 def test_MultilevelTetrahedralMesh():
     n = 2
@@ -619,3 +630,167 @@ def test_MultilevelTetrahedralMesh():
             ok_((elementChildren[l] == mlMesh.elementChildrenArrayList[l]).all())
             ok_((elementChildrenOffsets[l] == mlMesh.elementChildrenOffsetsList[l]).all())
         ok_((elementParents[l] == mlMesh.elementParentsArrayList[l]).all())
+    mlMesh2 = MultilevelTetrahedralMesh(0,0,0,skipInit=True)
+    mlMesh2.generateFromExistingCoarseMesh(mlMesh.meshList[0],
+                                           refinementLevels=n)
+
+def test_MultilevelHexahedralMesh():
+    n = 1
+    mlMesh = MultilevelHexahedralMesh(4,4,4,
+                                      refinementLevels=n)
+
+def setupStepGauss():
+    import numpy as np
+    from math import sin,cos,pi,sqrt,exp
+    #set up a fake LiDAR point set
+    nPoints_x = nPoints_y = 21
+    delta_x = 2.0/float(nPoints_x-1)
+    delta_y = 2.0/float(nPoints_y-1)
+    bathy = np.zeros((nPoints_x*nPoints_y,3),'d')
+    for i in range(nPoints_y):
+        for j in range(nPoints_x):
+            x = -0.5+j*delta_x
+            y = -0.5+i*delta_y
+            #z = 1.0
+            if y > x:
+                z = 1.0
+            else:
+                if y < x - 0.25:
+                    r = sqrt((y - 0.25)**2 + (x - 0.8)**2)
+                    z = exp(-50.0*r**2)
+                else:
+                    z = 0.0
+            #z = y
+            #z += sin(2.0*pi*x)*cos(2.0*pi*y)
+            bathy[i*nPoints_x+j,0] = x
+            bathy[i*nPoints_x+j,1] = y
+            bathy[i*nPoints_x+j,2] = z
+    domain = InterpolatedBathymetryDomain(vertices=[[0.0,0.0],
+                                                    [0.0,1.0],
+                                                    [0.5,1.5],
+                                                    [1.0,1.0],
+                                                    [1.5,-0.5]],
+                                          vertexFlags=[1,2,3,2,1],
+                                          segments=[[0,1],
+                                                    [1,2],
+                                                    [2,3],
+                                                    [3,4],
+                                                    [4,0]],
+                                          segmentFlags=[1,2,3,3,1],
+                                          regions=[(0.5,0.5)],
+                                          regionFlags=[1],
+                                          name="interpolatedBathySimpleTest",
+                                          units='m',
+                                          bathy = bathy,
+                                          bathyGridDim = (nPoints_y,nPoints_x))
+    domain.writePoly(domain.name)
+    return domain
+
+def test_L1():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="points",bathyAssignmentScheme="localAveraging",errorNormType="L1")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_L1_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+
+def test_L2():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="points",bathyAssignmentScheme="localAveraging",errorNormType="L2")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_L2_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+
+def test_Linfty():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="points",bathyAssignmentScheme="localAveraging",errorNormType="Linfty")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_Linfty_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+
+def test_L1_interp():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="points",bathyAssignmentScheme="interpolation",errorNormType="L1")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_L1_interp_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+
+def test_L2_interp():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="points",bathyAssignmentScheme="interpolation",errorNormType="L2")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_L2_interp_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+
+def test_Linfty_interp():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="points",bathyAssignmentScheme="interpolation",errorNormType="Linfty")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_Linfty_interp_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+
+def test_L1_grid():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="grid",bathyAssignmentScheme="localAveraging",errorNormType="L1")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_grid_L1_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+
+def test_L2_grid():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="grid",bathyAssignmentScheme="localAveraging",errorNormType="L2")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_grid_L2_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+
+def test_Linfty_grid():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="grid",bathyAssignmentScheme="localAveraging",errorNormType="Linfty")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_grid_Linfty_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+
+def test_L1_interp_grid():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="grid",bathyAssignmentScheme="interpolation",errorNormType="L1")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_grid_L1_interp_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+
+def test_L2_interp_grid():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="grid",bathyAssignmentScheme="interpolation",errorNormType="L2")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_grid_L2_interp_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+
+def test_Linfty_interp_grid():
+    domain = setupStepGauss()
+    mesh = InterpolatedBathymetryMesh(domain,triangleOptions="gVApq30Dena%8.8f" % (0.5**3,),atol=1.0e-1,rtol=1.0e-1,maxLevels=25,maxNodes=50000,
+                                      bathyType="grid",bathyAssignmentScheme="interpolation",errorNormType="Linfty")
+    archive = XdmfArchive(dataDir='.',filename="interpolatedBathySimpleTest_grid_Linfty_interp_")
+    archive.domain = ElementTree.SubElement(archive.tree.getroot(),"Domain")
+    mesh.meshList[-1].writeMeshXdmf(ar=archive,init=True)
+    archive.sync(); archive.close()
+    mesh.meshList[-1].writeMeshADH("interpolatedBathySimpleTest_grid_Linfty_interp_")
